+++ skrl/agents/jax/sac/sac.py
@@ -27,7 +27,7 @@ SAC_DEFAULT_CONFIG = {
 
     "actor_learning_rate": 1e-3,    # actor learning rate
     "critic_learning_rate": 1e-3,   # critic learning rate
-    "learning_rate_scheduler": None,        # learning rate scheduler class (see torch.optim.lr_scheduler)
+    "learning_rate_scheduler": None,        # learning rate scheduler function (see optax.schedules)
     "learning_rate_scheduler_kwargs": {},   # learning rate scheduler's kwargs (e.g. {"step_size": 1e-3})
 
     "state_preprocessor": None,             # state preprocessor class (see skrl.resources.preprocessors)
@@ -262,25 +262,29 @@ class SAC(Agent):
 
         # set up optimizers and learning rate schedulers
         if self.policy is not None and self.critic_1 is not None and self.critic_2 is not None:
+            # schedulers
+            if self._learning_rate_scheduler:
+                self.policy_scheduler = self._learning_rate_scheduler(**self.cfg["learning_rate_scheduler_kwargs"])
+                self.critic_scheduler = self._learning_rate_scheduler(**self.cfg["learning_rate_scheduler_kwargs"])
+            # optimizers
             with jax.default_device(self.device):
                 self.policy_optimizer = Adam(
-                    model=self.policy, lr=self._actor_learning_rate, grad_norm_clip=self._grad_norm_clip
+                    model=self.policy,
+                    lr=self._actor_learning_rate,
+                    grad_norm_clip=self._grad_norm_clip,
+                    scale=not self._learning_rate_scheduler,
                 )
                 self.critic_1_optimizer = Adam(
-                    model=self.critic_1, lr=self._critic_learning_rate, grad_norm_clip=self._grad_norm_clip
+                    model=self.critic_1,
+                    lr=self._critic_learning_rate,
+                    grad_norm_clip=self._grad_norm_clip,
+                    scale=not self._learning_rate_scheduler,
                 )
                 self.critic_2_optimizer = Adam(
-                    model=self.critic_2, lr=self._critic_learning_rate, grad_norm_clip=self._grad_norm_clip
-                )
-            if self._learning_rate_scheduler is not None:
-                self.policy_scheduler = self._learning_rate_scheduler(
-                    self.policy_optimizer, **self.cfg["learning_rate_scheduler_kwargs"]
-                )
-                self.critic_1_scheduler = self._learning_rate_scheduler(
-                    self.critic_1_optimizer, **self.cfg["learning_rate_scheduler_kwargs"]
-                )
-                self.critic_2_scheduler = self._learning_rate_scheduler(
-                    self.critic_2_optimizer, **self.cfg["learning_rate_scheduler_kwargs"]
+                    model=self.critic_2,
+                    lr=self._critic_learning_rate,
+                    grad_norm_clip=self._grad_norm_clip,
+                    scale=not self._learning_rate_scheduler,
                 )
 
             self.checkpoint_modules["policy_optimizer"] = self.policy_optimizer
@@ -490,8 +494,12 @@ class SAC(Agent):
             # optimization step (critic)
             if config.jax.is_distributed:
                 grad = self.critic_1.reduce_parameters(grad)
-            self.critic_1_optimizer = self.critic_1_optimizer.step(grad, self.critic_1)
-            self.critic_2_optimizer = self.critic_2_optimizer.step(grad, self.critic_2)
+            self.critic_1_optimizer = self.critic_1_optimizer.step(
+                grad, self.critic_1, self._critic_learning_rate if self._learning_rate_scheduler else None
+            )
+            self.critic_2_optimizer = self.critic_2_optimizer.step(
+                grad, self.critic_2, self._critic_learning_rate if self._learning_rate_scheduler else None
+            )
 
             # compute policy (actor) loss
             grad, policy_loss, log_prob = _update_policy(
@@ -508,7 +516,9 @@ class SAC(Agent):
             # optimization step (policy)
             if config.jax.is_distributed:
                 grad = self.policy.reduce_parameters(grad)
-            self.policy_optimizer = self.policy_optimizer.step(grad, self.policy)
+            self.policy_optimizer = self.policy_optimizer.step(
+                grad, self.policy, self._actor_learning_rate if self._learning_rate_scheduler else None
+            )
 
             # entropy learning
             if self._learn_entropy:
@@ -529,9 +539,8 @@ class SAC(Agent):
 
             # update learning rate
             if self._learning_rate_scheduler:
-                self.policy_scheduler.step()
-                self.critic_1_scheduler.step()
-                self.critic_2_scheduler.step()
+                self._actor_learning_rate *= self.policy_scheduler(timestep)
+                self._critic_learning_rate *= self.critic_scheduler(timestep)
 
             # record data
             if self.write_interval > 0:
@@ -555,6 +564,5 @@ class SAC(Agent):
                     self.track_data("Coefficient / Entropy coefficient", self._entropy_coefficient.item())
 
                 if self._learning_rate_scheduler:
-                    self.track_data("Learning / Policy learning rate", self.policy_scheduler.get_last_lr()[0])
-                    self.track_data("Learning / Critic 1 learning rate", self.critic_1_scheduler.get_last_lr()[0])
-                    self.track_data("Learning / Critic 2 learning rate", self.critic_2_scheduler.get_last_lr()[0])
+                    self.track_data("Learning / Policy learning rate", self._actor_learning_rate)
+                    self.track_data("Learning / Critic learning rate", self._critic_learning_rate)