+++ skrl/agents/jax/td3/td3.py
@@ -26,7 +26,7 @@ TD3_DEFAULT_CONFIG = {
 
     "actor_learning_rate": 1e-3,    # actor learning rate
     "critic_learning_rate": 1e-3,   # critic learning rate
-    "learning_rate_scheduler": None,        # learning rate scheduler class (see torch.optim.lr_scheduler)
+    "learning_rate_scheduler": None,        # learning rate scheduler function (see optax.schedules)
     "learning_rate_scheduler_kwargs": {},   # learning rate scheduler's kwargs (e.g. {"step_size": 1e-3})
 
     "state_preprocessor": None,             # state preprocessor class (see skrl.resources.preprocessors)
@@ -243,25 +243,29 @@ class TD3(Agent):
 
         # set up optimizers and learning rate schedulers
         if self.policy is not None and self.critic_1 is not None and self.critic_2 is not None:
+            # schedulers
+            if self._learning_rate_scheduler is not None:
+                self.policy_scheduler = self._learning_rate_scheduler(**self.cfg["learning_rate_scheduler_kwargs"])
+                self.critic_scheduler = self._learning_rate_scheduler(**self.cfg["learning_rate_scheduler_kwargs"])
+            # optimizers
             with jax.default_device(self.device):
                 self.policy_optimizer = Adam(
-                    model=self.policy, lr=self._actor_learning_rate, grad_norm_clip=self._grad_norm_clip
+                    model=self.policy,
+                    lr=self._actor_learning_rate,
+                    grad_norm_clip=self._grad_norm_clip,
+                    scale=not self._learning_rate_scheduler,
                 )
                 self.critic_1_optimizer = Adam(
-                    model=self.critic_1, lr=self._critic_learning_rate, grad_norm_clip=self._grad_norm_clip
+                    model=self.critic_1,
+                    lr=self._critic_learning_rate,
+                    grad_norm_clip=self._grad_norm_clip,
+                    scale=not self._learning_rate_scheduler,
                 )
                 self.critic_2_optimizer = Adam(
-                    model=self.critic_2, lr=self._critic_learning_rate, grad_norm_clip=self._grad_norm_clip
-                )
-            if self._learning_rate_scheduler is not None:
-                self.policy_scheduler = self._learning_rate_scheduler(
-                    self.policy_optimizer, **self.cfg["learning_rate_scheduler_kwargs"]
-                )
-                self.critic_1_scheduler = self._learning_rate_scheduler(
-                    self.critic_1_optimizer, **self.cfg["learning_rate_scheduler_kwargs"]
-                )
-                self.critic_2_scheduler = self._learning_rate_scheduler(
-                    self.critic_2_optimizer, **self.cfg["learning_rate_scheduler_kwargs"]
+                    model=self.critic_2,
+                    lr=self._critic_learning_rate,
+                    grad_norm_clip=self._grad_norm_clip,
+                    scale=not self._learning_rate_scheduler,
                 )
 
             self.checkpoint_modules["policy_optimizer"] = self.policy_optimizer
@@ -534,8 +538,12 @@ class TD3(Agent):
             # optimization step (critic)
             if config.jax.is_distributed:
                 grad = self.critic_1.reduce_parameters(grad)
-            self.critic_1_optimizer = self.critic_1_optimizer.step(grad, self.critic_1)
-            self.critic_2_optimizer = self.critic_2_optimizer.step(grad, self.critic_2)
+            self.critic_1_optimizer = self.critic_1_optimizer.step(
+                grad, self.critic_1, self._critic_learning_rate if self._learning_rate_scheduler else None
+            )
+            self.critic_2_optimizer = self.critic_2_optimizer.step(
+                grad, self.critic_2, self._critic_learning_rate if self._learning_rate_scheduler else None
+            )
 
             # delayed update
             self._critic_update_counter += 1
@@ -549,7 +557,9 @@ class TD3(Agent):
                 # optimization step (policy)
                 if config.jax.is_distributed:
                     grad = self.policy.reduce_parameters(grad)
-                self.policy_optimizer = self.policy_optimizer.step(grad, self.policy)
+                self.policy_optimizer = self.policy_optimizer.step(
+                    grad, self.policy, self._actor_learning_rate if self._learning_rate_scheduler else None
+                )
 
                 # update target networks
                 self.target_policy.update_parameters(self.policy, polyak=self._polyak)
@@ -558,9 +568,8 @@ class TD3(Agent):
 
             # update learning rate
             if self._learning_rate_scheduler:
-                self.policy_scheduler.step()
-                self.critic_1_scheduler.step()
-                self.critic_2_scheduler.step()
+                self._actor_learning_rate *= self.policy_scheduler(timestep)
+                self._critic_learning_rate *= self.critic_scheduler(timestep)
 
             # record data
             if not self._critic_update_counter % self._policy_delay:
@@ -580,6 +589,5 @@ class TD3(Agent):
             self.track_data("Target / Target (mean)", target_values.mean().item())
 
             if self._learning_rate_scheduler:
-                self.track_data("Learning / Policy learning rate", self.policy_scheduler.get_last_lr()[0])
-                self.track_data("Learning / Critic 1 learning rate", self.critic_1_scheduler.get_last_lr()[0])
-                self.track_data("Learning / Critic 2 learning rate", self.critic_2_scheduler.get_last_lr()[0])
+                self.track_data("Learning / Policy learning rate", self._actor_learning_rate)
+                self.track_data("Learning / Critic learning rate", self._critic_learning_rate)